Add issue prev/next navigation when issue is opened from issue query page

+++ core/src/main/java/io/onedev/server/manager/IssueManager.java
@@ -29,7 +29,7 @@ public interface IssueManager extends EntityManager<Issue> {
 	
 	List<Issue> query(IssueQuery issueQuery, int firstResult, int maxResults);
 	
-	long count(@Nullable IssueCriteria issueCriteria);
+	int count(@Nullable IssueCriteria issueCriteria);
 	
 	List<Issue> query(Project project, @Nullable String term, int count);
